refactor: split  and eliminate  clone per request
