+++ crates/agentic-core/benches/executor_throughput.rs
@@ -153,13 +153,7 @@ fn build_exec_ctx(rt: &tokio::runtime::Runtime, mock_url: String) -> (Arc<Execut
     let conv_handler = ConversationHandler::new(ConversationStore::new(pool.clone()));
     let resp_handler = ResponseHandler::new(ResponseStore::new(pool.clone()));
     let client = Arc::new(reqwest::Client::new());
-    let exec_ctx = Arc::new(ExecutionContext::new(
-        conv_handler,
-        resp_handler,
-        client,
-        mock_url,
-        None,
-    ));
+    let exec_ctx = Arc::new(ExecutionContext::new(conv_handler, resp_handler, client, mock_url));
     (exec_ctx, pool)
 }
